### 2021-08-27 #3663 @whedon
[PRE REVIEW]: AuDoLab: Automatic document labelling and classification for extremely unbalanced data
**Submitting author:** @ArneTillmann (<a href="http://orcid.org/0000-0002-2761-4518">Arne Matthias Tillmann</a>)
**Repository:** <a href="https://github.com/ArneTillmann/AuDoLab" target ="_blank">https://github.com/ArneTillmann/AuDoLab</a>
**Version:** v1.0.7
