### 2021-08-28 #3663 @ArneTillmann
👋 @danielskatz,

> 👋 @ArneTillmann - Your paper header includes AAS info. Is this associated with a AAS paper? If not, please remove that part of the paper source.
>
> In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.

Ofc I will make those changes to the paper asap.

Thank you very much
