### 2021-08-31 #3663 @arfon
:wave: @ArneTillmann I will be the editor for this paper. Thanks so much for your suggestions for reviewers!

While we wait to hear back from possible reviewers could you merge this PR with a few minor changes to your paper here: https://github.com/ArneTillmann/AuDoLab/pull/31
