### 2021-09-08 #3663 @arfon
:wave: @UTH-Tuan – you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [_AuDoLab: Automatic document labelling and1 classification for extremely unbalanced data_](https://github.com/ArneTillmann/AuDoLab).
