### 2021-09-08 #3663 @arfon
> desilinguist, sjvrijn, alexhanna, ahurriyetoglu and UTH-Tuan

BTW, most of these folks are already actively reviewing a JOSS submission so I don't want to ask them to review a second paper at this time. Thanks for the suggestions though!
