### 2021-09-11 #3719 @whedon
[REVIEW]: AuDoLab: Automatic document labelling and1 classfication for extremely unbalanced data
**Submitting author:** @ArneTillmann (<a href="http://orcid.org/0000-0002-2761-4518">Arne Matthias Tillmann</a>)
**Repository:** <a href="https://github.com/ArneTillmann/AuDoLab" target ="_blank">https://github.com/ArneTillmann/AuDoLab</a>
**Version:** v1.0.7
