### 2021-09-11 #3719 @arfon
@linuxscout – This is the review thread for the paper. All of our communications will happen here from now on.
