### 2021-09-12 #3719 @ChrisW09
@arfon fantastic, thank you! @linuxscout  @pps121 should you have any early ideas or comments, on how to improve the paper or make things more clear, please just let us know, and we will implement them as soon as we can. thank a lot.
