### 2021-09-17 ArneTillmann/AuDoLab#32 @linuxscout
Improve description
Hi,
I think that the desciption in README file is too short.
https://github.com/openjournals/joss-reviews/issues/3719

- [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? ** No**

Thanks
