### 2021-10-10 #3719 @pps121
When I ran the code:  from nltk.corpus import reuters

it gave me LookupError as below :
**********************************************************************
  Resource reuters not found.
  Please use the NLTK Downloader to obtain the resource:

  >>> import nltk
  >>> nltk.download('reuters')
