### 2021-10-10 #3719 @pps121
For the above LookUpError Below code gives a fix:

>>import nltk
>>nltk.download('reuters')

It downloads a reuters zip inside /nltk_data/corpora, then reloads to execute the cell will proceed to next step.
