### 2021-10-10 #3719 @pps121
Before executing the line
preprocessed_target = audo.text_cleaning(data=data, column="text")

We need to load wordnet as below otherwise there will be a module not found runtime error
>>import nltk
>>nltk.download('wordnet')
