### 2021-10-13 #3719 @ArneTillmann
>
>
> Before executing the line preprocessed_target = audo.text_cleaning(data=data, column="text")
>
> We need to load wordnet as below otherwise there will be a module not found runtime error
>
> > > import nltk
> > > nltk.download('wordnet')

as well as those.
