+++ .travis.yml
@@ -9,7 +9,7 @@ deploy:
   distributions: sdist bdist_wheel
   username: ArneTillmann
   password:
-    secure: N+43r/2dkFEuH6tJaz+6OkTeJmPJbTTLeLfL1JqG8akNHCqtDjozKVueUTADOTxFvYJJf4tEYnt8r389CuzM/mpm4afL72AHxOdkeEVlEs9oFe1J59qmPDahPtgcGjiqXpSNWe6BfSxPa0MflX05MismxpzOZF/GvOjdyWc4hkcT97n4L/cG0wrfE4segOi02QcO1UmDOQ0jcT1haqL3Y8tQfIkOz3kAJTkWusEUjGC0eJaahzot2Biai1W8kG8a+jTYg67n/wz4oydWI8mWlSxvECYP2vndsTK9kare0m1fIQkX8qcawCHvtOtS8gSpisxVZLER1f+0+SUbYAQgx8tfj9WluWj36URWGSECYD3z8/6S203QnMDA4B5qk4pvJ2tyFvp5ks53VEdAXbp30P6d/Hz6RHn4t1SIjiIX6LcdotYr/ZC5BIhdnhq33mhcBjfj1K+An0u9zGVKPyEEBakZD94sWF2qI+kfKvVVU1e3lnMKDpEjikyS5rk8FZhZpUX4Q3pGKXQKUYIIcQc4+OVfw+wOF/iqbI+1CJgRfokV2VpXIUG5wEGnjtlmXG08omj7WRhhFkx4X/8AWm+N0UiDFoeyjeVG98RyvuQ8OqtSfNxA8unViKX1zVoggoTiHUq0aZ52nlEPHtJPqH8HRN9q8T/wPpMjjd5EXGp2Owg=
+    secure: AXws5qxlusuAu+MCw9gI1osr0RAGb2dPTq/clG1BwVU5wK90qKcXANEOpZxYLhKlu1BVyWCAOXpXcOhiEvdWSzfG3LLRvHeIAZHI5ZdqKQeUt7/6JhFjjKIR2acYkkG+Q7RH4Nvih/c+szByex1AoWp1Ux2Wr/DmgLdCneFUVSQPWPzbxPjGQx4k5cV1tWBkfZD1qe8NjHbjylgRaEIDdXszObqUJGBhrnZOlA06PxkjOfnQ5A9kSbuLIqtNLXd/GNcgIcor0jGyJkvOke1Hy6I0m2FedpRL5TvZgCEgKkst6wkh/yKqEghXhTvNqUQBp20XCqNx9t2McoRiTZ7LeWOictGE2YN6OXs3j7VBSkc2T5iWEmCxMlsss/4K3jrb3V8yVqE7OHfzPIgEvuaugwyFK5o7KsKF1GOnjimPbv0ZTovRXDrS5AcrCyQ1yHIaVfBcP4Msn+mYe2Kj3qPYwGnqRO3F0TMeems/hZNuulzRxVb114mLi2qsF6ERTvp2uxqowVe/1MMGxNpsn+wMkJ/bmh1IvNb2wH3kasG0jqV19MiwutFylQBF6Zf95nFYn92b4UQsVxKR3nQ4DtJWA0YxulChYoic7RnC78LpgY2EvueX39S3MiEbwAE7HAKDiTqA8M7voMzykIeHkJWvERpUuf5kif8eHrPYR8v4+Cg=
   on:
     tag: true
     branch: main
