+++ setup.py
@@ -47,6 +47,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/ArneTillmann/AuDoLab",
-    version='1.0.9',
+    version='1.0.10',
     zip_safe=False,
 )
