+++ .travis.yml
@@ -9,7 +9,7 @@ deploy:
   distributions: sdist bdist_wheel
   username: ArneTillmann
   password:
-    secure: AXws5qxlusuAu+MCw9gI1osr0RAGb2dPTq/clG1BwVU5wK90qKcXANEOpZxYLhKlu1BVyWCAOXpXcOhiEvdWSzfG3LLRvHeIAZHI5ZdqKQeUt7/6JhFjjKIR2acYkkG+Q7RH4Nvih/c+szByex1AoWp1Ux2Wr/DmgLdCneFUVSQPWPzbxPjGQx4k5cV1tWBkfZD1qe8NjHbjylgRaEIDdXszObqUJGBhrnZOlA06PxkjOfnQ5A9kSbuLIqtNLXd/GNcgIcor0jGyJkvOke1Hy6I0m2FedpRL5TvZgCEgKkst6wkh/yKqEghXhTvNqUQBp20XCqNx9t2McoRiTZ7LeWOictGE2YN6OXs3j7VBSkc2T5iWEmCxMlsss/4K3jrb3V8yVqE7OHfzPIgEvuaugwyFK5o7KsKF1GOnjimPbv0ZTovRXDrS5AcrCyQ1yHIaVfBcP4Msn+mYe2Kj3qPYwGnqRO3F0TMeems/hZNuulzRxVb114mLi2qsF6ERTvp2uxqowVe/1MMGxNpsn+wMkJ/bmh1IvNb2wH3kasG0jqV19MiwutFylQBF6Zf95nFYn92b4UQsVxKR3nQ4DtJWA0YxulChYoic7RnC78LpgY2EvueX39S3MiEbwAE7HAKDiTqA8M7voMzykIeHkJWvERpUuf5kif8eHrPYR8v4+Cg=
+    secure: eyYZlfGrvhkKKCVhCcWDihbHKldl3lNsbXsPbQIPBdN2aRoeQN5baIgKImxZKusNpkYjdLYwBfLM394jjOkur814b/ga7lf8eIq4o8R5GM9+Msdsbx5irdCwNXqoPijuaIRP90LIiBVHS1g2FOVNef3Cvd/eW9W2NgMcMsSptalCfU98/B4fFVfF7efgK4MOxoGd/97g+0I9/eake6Z/10D6Ukg9hcOE2MfSt9f+j5ydhNptSviiyneioAcBuPOgFRHvR+Cjgk9nyUTvf7I3Nh3CnuLnRzBemT0euzZTUpwh0G+wJQ3gLW2ePDiCz/8H4YHTOT8t29vGcRRHhwKgG3AyxWttaEOzEHVZTQPnSbvWtSkNTZV0HDiF29+5oCGAfv5dsUUg9dLGooRHhNYPbvwzmzvvZIgjjy/5DzzqSF9b9oScXHxILqTxUj43yvgJIUoycckyR8S+1lfDfwwbO9atDbn9bVRXnNJwXFrPPiEpohM/8KlsW2exnoo0xF3dCgY1NuUFr0A9W7y1AWQzcfgb9VUTVhwu8smvWUMCtKQ5DCZO0u1t7U0qUGrgPC+CIAhRuOXwocTRdGTurn1/FRbrYdHA1hjk3AmFyhvNLFpwJKktH084VvQtdguX6TD67BnYxkkBkllVEc5TkzK1NNYwjJJ1OFOMZWMvyhyog88=
   on:
     tag: true
     branch: main
