+++ .travis.yml
@@ -7,9 +7,9 @@ deploy:
   skip_cleanup: true
   provider: pypi
   distributions: sdist bdist_wheel
-  username: ArneTillmann
+  username: __token__
   password:
-    secure: eyYZlfGrvhkKKCVhCcWDihbHKldl3lNsbXsPbQIPBdN2aRoeQN5baIgKImxZKusNpkYjdLYwBfLM394jjOkur814b/ga7lf8eIq4o8R5GM9+Msdsbx5irdCwNXqoPijuaIRP90LIiBVHS1g2FOVNef3Cvd/eW9W2NgMcMsSptalCfU98/B4fFVfF7efgK4MOxoGd/97g+0I9/eake6Z/10D6Ukg9hcOE2MfSt9f+j5ydhNptSviiyneioAcBuPOgFRHvR+Cjgk9nyUTvf7I3Nh3CnuLnRzBemT0euzZTUpwh0G+wJQ3gLW2ePDiCz/8H4YHTOT8t29vGcRRHhwKgG3AyxWttaEOzEHVZTQPnSbvWtSkNTZV0HDiF29+5oCGAfv5dsUUg9dLGooRHhNYPbvwzmzvvZIgjjy/5DzzqSF9b9oScXHxILqTxUj43yvgJIUoycckyR8S+1lfDfwwbO9atDbn9bVRXnNJwXFrPPiEpohM/8KlsW2exnoo0xF3dCgY1NuUFr0A9W7y1AWQzcfgb9VUTVhwu8smvWUMCtKQ5DCZO0u1t7U0qUGrgPC+CIAhRuOXwocTRdGTurn1/FRbrYdHA1hjk3AmFyhvNLFpwJKktH084VvQtdguX6TD67BnYxkkBkllVEc5TkzK1NNYwjJJ1OFOMZWMvyhyog88=
+    secure: T0BeO0yhmB0DTBtSC3kmwQS7OXE2g+E5XgqWckMF849/wCIKQbxlhiyH4Nje9PiFZGjQE5RcccKE+n+X4709Irz42PvIe3XvE40Y/FwAnTp3q/gKXJDqnUt2RBQpZ2Mt0Pf8P9rDS8Qk/tYgWD7gMLlkPlLFizqLJhoNwpjsgo3yb+Ga6NmxLX0FRmmDSVO9QhD+Je3UHNsiJ0hjkApL9kltcMxVMIwIeK7+ZkH3FSId7iHaMFx9sgtfbAHG2sZ0KMPcgyzJEDcabIeDjZW2+Lm+x2mfJbszctObYPlxSvvN4h2ziWQBq0hQnAVRBoAeDExO/6r6+2PUl/Exgi6wZiq4tCGX7q8W9IGVYlvxMEeFNsP4+4NXaDvEJNCbDilOsICPdF5u0vp+I796A2205s7BcCcGxUsyOW92LlW9dfuD2YEY1RLDnIiYP6tEK4EAzwSKG03HhzfdLASN1cQmIZLiKc1A3B1sj/NL2A2svuToHWFUJIeTlz0X2hHGLNWhxnPE4wZG0kB/p5kPGZXuPb93+/xGnBMMZzjV1JIydRb4pMNoD51EllxTekWWDYFCqNzqNgS1L41dn+bknr883x9ySeEswM9XkW6q7KEqxPdCm256ScpnYTnnb+fOzuXLb1zV6gz8qburak2jjRCbtOBaJCLJCMi4NqdUyCUmDz8=
   on:
     tag: true
     branch: main
