### 2021-10-14 #3719 @pps121
> > The naming convention of ipython notebook should be more meaningful rather than example.ipynb
>
> I changed the name to usage_example.ipynb. Is that better, or do you have something else in mind?

It it better now. Thanks.
