### 2021-10-14 #3719 @pps121
> > The naming convention of ipython notebook should be more meaningful rather than example.ipynb
>
> I changed the name to usage_example.ipynb. Is that better, or do you have something else in mind?

``

> > ```
> > * The code level steps should be more elaborated, so that before executing it can give a high level purpose of its use.
> >
> > * In .travis.yml, the yaml expose a severe security hole: the username and password are mentioned as below:
> >   **username**: ArneTillmann
> >   **password**: 2AQeUe5iHHEe0MrEv7
> > ```
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> > Gitignore should take care of it before merging.
>
> Thank you especially for this hint. I didn't notice, but I changed the password and encrypted it now. However, I don't really know which files or lines of code you think should be more elaborated. Can you specify what you would prefer here?

Thanks for the encryption step. It is better now. it is okay now about the current state of the code modules.
