### 2021-10-15 dd0c8d1
update with examples and use cases
