### 2021-10-15 #3719 @ChrisW09
@pps121 Regarding your second comment "For different OS variants, can you include steps for both bash and PowerShell so that general users with different OS can find this more useful?".

We tested our package on various OS variants, namely macOS, unix and windows.

We will add explanations for the installation for both bash and PowerShell. This will make the package more useful for the general user as you suggested. Thank you for these suggestions!
