### 2021-10-15 0f2cbc8
For different OS variants, can you include steps for both bash and PowerShell so that general users with different OS can find this more useful?
