+++ paper.bib
@@ -139,12 +139,16 @@ doi = {10.17875/gup2021-1608},
 publisher = {Universitätsverlag Göttingen},
 }
 
-
-
-
-
-
-
-
-
-
+@Inbook{Thielmann2021,
+author={Thielmann, Anton
+and Weisser, Christoph
+and Krenz, Astrid},
+editor={Phuong, Nguyen Hoang
+and Kreinovich, Vladik},
+title={One-Class Support Vector Machine and LDA Topic Model Integration---Evidence for AI Patents},
+bookTitle={Soft Computing: Biomedical and Related Applications},
+year={2021},
+publisher={Springer International Publishing},
+pages={263--272},
+doi={10.1007/978-3-030-76620-7_23},
+}
