+++ paper.bib
@@ -127,16 +127,14 @@ author={{IEEE Xplore}},
  doi = {10.17875/gup2020-1338},
 }
 
-
 @book{Saefken2021,
-editor = {Kruse, René-Marcel},
-editor = {Säfken, Benjamin},
-editor = {Silbersdorff, Alexander},
-editor = {Weisser, Christoph},
-title = {Learning Deep Textwork: Perspectives on Natural Language Processing and Artificial Intelligence},
-year = {2021},
-doi = {10.17875/gup2021-1608},
-publisher = {Universitätsverlag Göttingen},
+ year = {2021},
+ title = {{Learning Deep Textwork: Perspectives on Natural Language Processing and Artificial Intelligence}},
+ address = {G{\"o}ttingen},
+ publisher = {{Universit{\"a}tsverlag G{\"o}ttingen}},
+ series = {{Universit{\"a}tsdrucke G{\"o}ttingen}},
+ editor = {S{\"a}fken, Benjamin and Silbersdorff, Alexander and Weisser, Christoph},
+ doi = {10.17875/gup2021-1608},
 }
 
 @Inbook{Thielmann2021,
