### 2021-10-19 #3719 @arfon
@ArneTillmann – could you take a look at the DOI suggestions from @whedon above? Also, `10.5555/1953048.2078195` doesn't resolve – can you check this reference?
