+++ paper.bib
@@ -98,7 +98,6 @@ doi = {10.1080/02664763.2021.1919063},
 
 
 @article{scikit-learn,
- doi = {10.5555/1953048.2078195},
  title={Scikit-learn: Machine Learning in {P}ython},
  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
          and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
