+++ paper.bib
@@ -109,11 +109,12 @@ doi = {10.1080/02664763.2021.1919063},
  year={2011}
 }
 
-@misc{IEEE,
-author={{IEEE Xplore}},
-  title={{IEEE Xplore Digital Library}},
-  year={2020},
-  url = {https://ieeexplore.ieee.org/Xplore/home.jsp}
+
+@webpage{IEEE,
+	Author = {{IEEE Xplore}},
+	Title = {{IEEE Xplore Digital Library}},
+	Url = {https://ieeexplore.ieee.org/Xplore/home.jsp},
+	Year = {2020},
 }
 
 @book{Saefken2020,
