### 2021-10-19 #3719 @arfon
@ArneTillmann @ChrisW09 – your paper is now accepted and published in JOSS :zap::rocket::boom:
