+++ README.rst
@@ -9,11 +9,11 @@ AuDoLab
         :target: https://travis-ci.com/ArneTillmann/AuDoLab
 
 .. image:: https://readthedocs.org/projects/audolab/badge/?version=latest
- :target: https://audolab.readthedocs.io/en/latest/?badge=latest
- :alt: Documentation Status
- 
- .. image:: https://joss.theoj.org/papers/10.21105/joss.03719/status.svg
-   :target: https://doi.org/10.21105/joss.03719
+        :target: https://audolab.readthedocs.io/en/latest/?badge=latest
+        :alt: Documentation Status
+
+.. image:: https://joss.theoj.org/papers/10.21105/joss.03719/status.svg
+        :target: https://doi.org/10.21105/joss.03719
 
 With AuDoLab you can perform Latend Direchlet Allocation on highly imbalanced datasets.
 
