### 2021-10-19 #3719 @arfon
> @arfon Somehow I still cannot see the PDF on the JOSS website. @ArneTillmann Can you see the PDF? thanks

@ChrisW09 – this is almost certainly your browser caching the old GitHub Pages response. I assure you this will automatically fix itself over the coming hours (and doesn't affect new visitors to the paper page)
